# src/main/java/com/example/rewardpointsdemo/model/CustomersBySum.java
package com.example.rewardpointsdemo.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CustomersBySum {

    private String customerName;
    private Integer rewardPoints;
}