# src/main/java/com/example/rewardpointsdemo/model/RewardPointsByMonth.java
package com.example.rewardpointsdemo.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.Month;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class RewardPointsByMonth {

    private Month month;
    private Integer amount;
}