# src/main/java/com/example/rewardpointsdemo/model/RewardPointsResponse.java
package com.example.rewardpointsdemo.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.io.Serializable;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class RewardPointsResponse implements Serializable {

    private Integer rewardPoints;
}