# src/main/java/com/example/rewardpointsdemo/model/RewardPointsSumResponse.java
package com.example.rewardpointsdemo.model;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class RewardPointsSumResponse {

    private List<CustomersBySum> customers;
}